requirements.txt
backports.statistics==0.1.0
colorama==0.3.7
flowiz==2.3.1
imageio==2.6.1
joblib==0.14.1
librosa==0.7.1
matplotlib==3.1.2
numpy==1.13.3
opencv_python==4.2.0.32
pandas==0.25.1
Pillow==7.0.0
pypesq==1.2.4
pystoi==0.2.2
pytube==9.5.2
pytz==2019.3
scikit_learn==0.22.2.post1
scipy==1.4.1
setproctitle==1.1.10
Shapely==1.6.4.post2
SoundFile==0.10.2
tensorboardX==2.0
torch==1.3.0
torchvision==0.4.1
tqdm==4.41.0